+++ docusaurus-search-local/src/client/utils/proxiedGeneratedConstants.ts
@@ -1,0 +1,2 @@
+// This file is auto generated while building.
+export * from "@generated/@easyops-cn/docusaurus-search-local/default/generated-constants.js";