+++ docusaurus-search-local/src/declarations.ts
@@ -3,16 +3,12 @@ declare module "@easyops-cn/autocomplete.js" {
 }
 
 declare module "*/generated.js" {
-  export const language: string[];
-  export const removeDefaultStopWordFilter: string[];
   export const removeDefaultStemmer: string[];
   export class Mark {
     constructor(root: HTMLElement);
     mark: (terms: string[], options?: Record<string, unknown>) => void;
     unmark: () => void;
   }
-  export const searchIndexUrl: string;
-  export const searchResultLimits: number;
   export const searchResultContextMaxLength: number;
   export const explicitSearchResultPath: boolean;
   export const searchBarShortcut: boolean;
@@ -27,10 +23,16 @@ declare module "*/generated.js" {
   export const hideSearchBarWithNoSearchContext: boolean;
   export const useAllContextsWithNoSearchContext: boolean;
   export const forceIgnoreNoIndex: boolean;
+}
+
+declare module "*/generated-constants.js" {
+  export const removeDefaultStopWordFilter: string[];
+  export const language: string[];
+  export const searchIndexUrl: string;
+  export const searchResultLimits: number;
   // These below are for mocking only.
   export const __setLanguage: (value: string[]) => void;
   export const __setRemoveDefaultStopWordFilter: (value: boolean) => void;
-  export const __setRemoveDefaultStemmer: (value: boolean) => void;
 }
 
 declare module "@docusaurus/Head";