+++ jest.config.js
@@ -5,10 +5,10 @@ module.exports = {
   },
   clearMocks: true,
   testMatch: ["**/docusaurus-search-local/src/**/?(*.)+(spec|test).[jt]s?(x)"],
-  testPathIgnorePatterns: ["/node_modules/", "/.yarn/", "<rootDir>/website/"],
+  testPathIgnorePatterns: ["/node_modules/", "/.yarn/", "<rootDir>/website/", "searchByWorker"],
   collectCoverage: true,
   coverageDirectory: "<rootDir>/.coverage",
-  collectCoverageFrom: ["docusaurus-search-local/src/**/*.{ts,tsx,js,jsx}"],
+  collectCoverageFrom: ["docusaurus-search-local/src/**/*.{ts,tsx,js,jsx}", "!docusaurus-search-local/src/**/searchByWorker.*"],
   moduleNameMapper: {
     "\\.module\\.css$": require.resolve("identity-obj-proxy"),
     "@docusaurus/Translate": "<rootDir>/__mocks__/Translate.js",